tests/wpt/web-platform-tests/html/semantics/embedded-content/the-img-element/null-image-source.html

<!doctype html>
<meta charset="utf-8">
<title></title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<img src="">
<script>
function check(img) {
    async_test(function(t) {
        img.onerror = t.step_func(function(e) {
            assert_equals(e.type, "error", "null image source check failed");
            t.done();
        });
    });
}
[].forEach.call(document.querySelectorAll('img'), check);
</script>
